+++ core/service/src/main/kotlin/io/mailit/core/service/id/InstanceIdProvider.kt
@@ -2,5 +2,5 @@ package io.mailit.core.service.id
 
 fun interface InstanceIdProvider {
 
-    fun getInstanceId(): Long
+    fun getInstanceId(): Int
 }
